LMS-24:Implemented get API for booking entity
