+++ src/main/java/com/java/librarymanagement/LibraryManagementApplication.java
@@ -7,6 +7,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 public class LibraryManagementApplication {
 
 	public static void main(String[] args) {
+
 		SpringApplication.run(LibraryManagementApplication.class, args);
 	}
 }
