+++ src/main/java/com/java/librarymanagement/booking/service/BookingServiceImpl.java
@@ -5,12 +5,10 @@ import com.java.librarymanagement.booking.model.Booking;
 import com.java.librarymanagement.booking.model.BookingDTO;
 import com.java.librarymanagement.booking.repository.BookingRepository;
 import com.java.librarymanagement.users.model.User;
-import com.java.librarymanagement.users.service.IUserService;
+import com.java.librarymanagement.users.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-import java.time.LocalDate;
+
 import java.util.List;
 
 @Service
@@ -20,49 +18,13 @@ public class BookingServiceImpl implements BookingService {
     private BookingRepository bookingRepository;
 
     @Autowired
-    private IUserService userService;
-    // Method to update overdue days and fine amounts for all bookings
-    public void updateOverdueAndFines(LocalDate currentDate) {
-        // Fetch all bookings where the due date is before the current date
-        List<Booking> bookings = bookingRepository.findByDueDateBefore(currentDate);
-
-        for (Booking booking : bookings) {
-            // Calculate overdue and fine for each booking
-            booking.calculateOverdueAndFine(currentDate);
-            // Save the updated booking with overdue and fine
-            bookingRepository.save(booking);
-        }
-    }
-    // Method to get all overdue books
-    public List<Booking> getAllOverdueBooks(LocalDate currentDate) {
-        return bookingRepository.findByDueDateBefore(currentDate);
-    }
-
-    // Method to get all overdue books with overdue days calculated
-    public List<Booking> getAllOverdueBooksWithDays(LocalDate currentDate) {
-        // Fetch all overdue bookings (dueDate before currentDate)
-        List<Booking> overdueBookings = bookingRepository.findByDueDateBefore(currentDate);
-
-        // For each booking, calculate the overdue days
-        overdueBookings.forEach(booking -> {
-            int overdueDays = calculateOverdueDays(booking.getDueDate(), currentDate);
-            booking.setOverDue(overdueDays);  // Set the overdue days
-        });
-
-        return overdueBookings;
-    }
-
-    // Helper method to calculate overdue days
-    private int calculateOverdueDays(LocalDate dueDate, LocalDate currentDate) {
-        if (dueDate != null && dueDate.isBefore(currentDate)) {
-            return (int) java.time.temporal.ChronoUnit.DAYS.between(dueDate, currentDate);
-        }
-        return 0;  // No overdue days
-    }
+    private UserService userService;
 
     @Override
     public List<BookingDTO> findAll() {
-        return List.of();
+
+        List<Booking> bookings= this.bookingRepository.findAll();
+        return BookingMapper.toDTO(bookings);
     }
 
     @Override
@@ -70,7 +32,7 @@ public class BookingServiceImpl implements BookingService {
         User authenticatedUser = userService.fetchSelfInfo();
         entity.setUser(authenticatedUser);
         Booking savedBooking = bookingRepository.save(entity);
-        Booking savedBooking = bookingRepository.save(entity);
+
         return BookingMapper.toDTO(savedBooking);
     }
