+++ src/main/java/com/java/librarymanagement/books/controller/BookController.java
@@ -91,55 +91,3 @@ public class BookController {
         return RestHelper.responseMessage(message);
     }
 }
-
-
-    /**
-     * Fetches all the user entities in the system.
-     *
-     * @return The list of user entities.
-     */
-    @GetMapping
-    @PreAuthorize("hasAuthority('ADMIN')")
-    public ResponseEntity<RestResponse> findAll() {
-        HashMap<String, Object> listHashMap = new HashMap<>();
-        listHashMap.put("book", bookService.findAll());
-        return RestHelper.responseSuccess(listHashMap);
-    }
-
-    /**
-     * Deletes the user by id.
-     *
-     * @param id The unique identifier of the entity.
-     * @return The message indicating the confirmation on deleted user entity.
-     */
-    @DeleteMapping("/{id}")
-    @PreAuthorize("hasAuthority('ADMIN')")
-    public ResponseEntity<RestResponse> delete(@PathVariable long id) {
-        String message = bookService.deleteById(id);
-        return RestHelper.responseMessage(message);
-    }
-
-    /**
-     * Counts the books in the book entity.
-     *
-     * @return The message indicating the total number of books in book entity.
-     */
-
-    @GetMapping("/count")
-    public long getTotalBooks() {
-        return bookRepository.count();
-    }
-
-    /**
-     * Updates the existing book entity.
-     *
-     * @param id The updated book entity.
-     * @return The message indicating the confirmation on updated book entity.
-     */
-    @PatchMapping("/update/{id}")
-    @PreAuthorize("hasAuthority('ADMIN')")
-    public ResponseEntity<RestResponse> update(@PathVariable long id, @RequestBody @Validated BookDTO bookDTO) {
-        String message = bookService.update(id, bookDTO);
-        return RestHelper.responseMessage(message);
-    }
-}
