+++ src/main/resources/application.properties
@@ -18,6 +18,7 @@ jwt.refresh-token-expiration=604800000
 # Cors Implementation
 cors.originUrl=
 
+
 # Email Setup
 spring.mail.host=smtp.gmail.com
 spring.mail.port=587
